refactor(type-designer): move sidebar wrapper into a dedicated folder and split special attribute
